+++ src/side_dialog/EntityView.tsx
@@ -25,7 +25,7 @@ import CloseIcon from "@material-ui/icons/Close";
 import { EntityPreview } from "../preview";
 import { EntityCollectionTable } from "../collection/EntityCollectionTable";
 import { removeInitialSlash } from "../routes/navigation";
-import { useSideEntityController } from "./SideEntityPanelsController";
+import { useSideEntityController } from "../contexts/SideEntityPanelsController";
 import CircularProgressCenter from "../components/CircularProgressCenter";